accept string slice for message
