+++ .rustfmt.toml
@@ -1,0 +1,3 @@
+imports_layout = "HorizontalVertical"
+imports_granularity = "Crate"
+max_width = 80
