+++ Cargo.toml
@@ -16,4 +16,5 @@ include = ["src/**/*", "README.md"]
 [dependencies]
 lazy_static = { version = "1.4.0" }
 maplit = { version = "1.0.2" }
+once_cell = "1.17.0"
 strum = { version = "0.24.0", features = ["derive"] }
