+++ examples/cycle.rs
@@ -5,7 +5,7 @@ use strum::IntoEnumIterator;
 fn main() {
     // loop through each spinner and display them for 2 seconds
     for spinner in Spinners::iter() {
-        let mut sp = Spinner::new(spinner.clone(), format!("{:?}", spinner));
+        let mut sp = Spinner::new(spinner.clone(), spinner.as_ref());
         sleep(Duration::from_secs(2));
         sp.stop();
     }
