+++ src/utils/mod.rs
@@ -1,4 +1,2 @@
-pub mod spinner_data;
-pub mod spinner_names;
-pub mod spinners_data;
+pub mod spinners;
 pub mod stream;
