--- src/utils/spinner_data.rs
@@ -1,5 +1,0 @@
-#[derive(Debug, Clone)]
-pub struct SpinnerData {
-    pub frames: Vec<&'static str>,
-    pub interval: u16,
-}
